LayoutTests/editing/deleting/delete-and-cleanup-expected.txt

This test checks that deletion does not leave unnecessary nested divs.

PASS confirmedMarkup is '<br>'
PASS confirmedMarkup is '<br>'
PASS confirmedMarkup is '<div id="mydiv"><br></div>'
PASS confirmedMarkup is '<br>'
PASS confirmedMarkup is '<b><i><br></i></b>'
PASS confirmedMarkup is '<b><div style="border: solid red"><i><br></i></div></b>'
PASS confirmedMarkup is '<br>'
PASS confirmedMarkup is '<br>'
PASS confirmedMarkup is '<p><br></p>'
PASS successfullyParsed is true

TEST COMPLETE
